retico_mistyrobot/retico_misty_camera_stream/requirements.txt
requests
numpy
av
opencv-python
pillow
matplotlib
retico_core @ git+https://github.com/retico-team/retico-core.git